Backend_ia/professional_ai_chat_agent/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sendgrid
google-generativeai
dateparser